Split scrape_boards.ts into topical modules
Now we're ready to rumble!
